pyproject.toml
[project]
name = "comet-examples-public"
version = "0.1.0"
description = ""
authors = [
    {name = "Francisco",email = "22344801+fschlz@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "comet-ml (>=3.49.11,<4.0.0)",
    "opik (>=1.7.43,<2.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "pandas (>=2.3.0,<3.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
